Finished working on docker
Everything should work now
Fixed frontend bugs
Fixed backend issues
Added in task generation to main project

+++ docker/dev-ob.sh
@@ -1,3 +1,5 @@
+#!/bin/bash
+
 source ./developer-observatory.conf
 
 RED='\033[1;31m'
@@ -14,18 +16,7 @@ prompt_confirm() {
   done  
 }
 
-
-if [[ -z $1 ]]; then
-    echo "Developer Observatory Control Script"
-    echo "Usage: dev-ob.sh [command]"
-    echo -e "\nAvailable commands:"
-    echo -e "install\t\t Installs all required packages and sets up services"
-    echo -e "configure\t Generates and complies configuration files"
-    echo -e "start\t\t Configures and starts the docker application"
-    echo -e "reset\t\t Completely resets the application to the inital state - ALL DATA IS DELETED"
-elif [[ $1 == "install" ]]; then
-    echo "Installing developer observatory"
-elif [[ $1 == "configure" ]]; then
+build_config() {
     if [[ -d "generator/generated" ]]; then
         echo "Generating configuration files and secrets"
         cp -r generator/generated containers/submit/tasks
@@ -35,29 +26,38 @@ elif [[ $1 == "configure" ]]; then
         exit 1
     fi
 
+    # Configuring secrets, since this has persistence
+    if [[ ! -f config/secrets ]]; then
+        cp config/Postgres.docker containers/postgres/Dockerfile
+        cp config/dbSchema.sql containers/postgres/
+        cp config/taskSchema.sql containers/postgres/
+        #Generate Passwords for the database in the first run. Replace in the files directly
+        pwUser0=$(cat /dev/urandom | tr -dc 'a-zA-Z0-9' | fold -w 32 | head -n 1)
+        pwUser1=$(cat /dev/urandom | tr -dc 'a-zA-Z0-9' | fold -w 32 | head -n 1)
+        pwUser2=$(cat /dev/urandom | tr -dc 'a-zA-Z0-9' | fold -w 32 | head -n 1)
+        pwUser3=$(cat /dev/urandom | tr -dc 'a-zA-Z0-9' | fold -w 32 | head -n 1)
+        sed -i "s|%pwUser0%|$pwUser0|g" containers/postgres/Dockerfile
+        sed -i "s|%pwUser1%|$pwUser1|g" containers/postgres/dbSchema.sql
+        sed -i "s|%pwUser2%|$pwUser2|g" containers/postgres/dbSchema.sql
+        sed -i "s|%pwUser3%|$pwUser3|g" containers/postgres/dbSchema.sql
+
+        echo -e "#/bin/bash\npwUser0=$pwUser0\npwUser1=$pwUser1\npwUser2=$pwUser2\npwUser3=$pwUser3" > config/secrets
+    else
+        source config/secrets
+    fi
+
+
     cp config/landing.php containers/landing/webpageConf/config.php
-    cp config/nginx.conf containers/nginx/
-    cp config/Postgres.docker containers/postgres/Dockerfile
-    cp config/dbSchema.sql containers/postgres/
-    cp config/taskSchema.sql containers/postgres/
     cp config/redis.conf containers/redis/
     cp config/submit.py containers/submit/configSubmit.py
     cp config/manager_config.py manager/
 
-    #Generate Passwords for the database in the first run. Replace in the files directly
-    pwUser0=$(cat /dev/urandom | tr -dc 'a-zA-Z0-9' | fold -w 32 | head -n 1)
-    pwUser1=$(cat /dev/urandom | tr -dc 'a-zA-Z0-9' | fold -w 32 | head -n 1)
-    pwUser2=$(cat /dev/urandom | tr -dc 'a-zA-Z0-9' | fold -w 32 | head -n 1)
-    pwUser3=$(cat /dev/urandom | tr -dc 'a-zA-Z0-9' | fold -w 32 | head -n 1)
-    sed -i "s|%pwUser0%|$pwUser0|g" containers/postgres/Dockerfile
-    sed -i "s|%pwUser1%|$pwUser1|g" containers/postgres/dbSchema.sql
-    sed -i "s|%pwUser2%|$pwUser2|g" containers/postgres/dbSchema.sql
-    sed -i "s|%pwUser3%|$pwUser3|g" containers/postgres/dbSchema.sql
 
     sed -i "s|%pwUser1%|$pwUser1|g" containers/landing/webpageConf/config.php
     sed -i "s|%finalSurveyURL%|$finalSurveyURL|g" containers/landing/webpageConf/config.php
 
     sed -i "s|%pwUser2%|$pwUser2|g" containers/submit/configSubmit.py
+    sed -i "s|%finalSurveyURL%|$finalSurveyURL|g" containers/submit/configSubmit.py
 
     sed -i "s|%poolSize%|$poolSize|g" manager/manager_config.py
 
@@ -83,8 +83,24 @@ elif [[ $1 == "configure" ]]; then
     cd instance/
     . configure_instance.sh
     cd ..
+}
+
+
+if [[ -z $1 ]]; then
+    echo "Developer Observatory Control Script"
+    echo "Usage: dev-ob.sh [command]"
+    echo -e "\nAvailable commands:"
+    echo -e "install\t\t Installs all required packages and sets up services"
+    echo -e "configure\t Generates configuration files"
+    echo -e "start\t\t Configures and starts the docker application"
+    echo -e "reset\t\t Completely resets the application to the inital state - ALL DATA IS DELETED"
+elif [[ $1 == "install" ]]; then
+    echo "Installing developer observatory"
+elif [[ $1 == "configure" ]]; then
+    build_config
 
 elif [[ $1 == "start" ]]; then
+    build_config
     docker-compose build && docker-compose up
 elif [[ $1 == "reset" ]]; then
     echo -e "${RED} WARNING: THIS WILL CLEAR ALL OF YOUR DATA, INCLUDING STUDY RESULTS${NC}"
@@ -93,12 +109,14 @@ elif [[ $1 == "reset" ]]; then
     # Run docker-compose down
     docker-compose down
 
+    # Clean secrets
+    rm config/secrets
+
     # Clean task files
     rm -r containers/submit/tasks
 
     # Clean generated configuration files
     rm containers/landing/webpageConf/config.php
-    rm containers/nginx/nginx.conf
     rm containers/postgres/*.sql
     rm containers/postgres/Dockerfile
     rm containers/submit/configSubmit.py
@@ -107,4 +125,7 @@ elif [[ $1 == "reset" ]]; then
     # Purge db volume
     docker volume rm docker_devob-data
 
+else
+    echo -e "${RED}Unknown command: $1"
+    exit 1
 fi
